+++ packages/components/atoms/f-form-field/CHANGELOG.md
@@ -4,12 +4,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v1.19.0
+------------------------------
+*July 22, 2021*
+
+### Added
+- `labelDetails` and `assistiveText`.
+- Styling to placeholder.
+
+
 v1.18.0
 ------------------------------
 *July 22, 2021*
 
 ### Added
-- Leading and Trailing icon slots.
+- Leading and Trailing con slots.
 - Tests to cover changes.